Unbeatable  Tic_Tac_Toe using Minimax algorithm
This project is basically a Tic_Tac_Toe game where a human player can play against an AI agent. The AI uses the Minimax algorithm to always make an optional move, making it unbeatable.
This is the rule based game as you see X is human and o is AI (computer) ,this game is 3*3 grid, win/draw conditions.
Human enter the moves and AI responds automatically with the best move.